🚚 move StoreTypes into store module & other misc cleanup
