+++ .github/workflows/ci-tests.yml
@@ -97,4 +97,4 @@ jobs:
     - run: make test-integration
     - name: Print logs
       if: failure()
-      run: docker service logs --tail 200 indra_node && docker service logs --tail 200 indra_proxy
+      run: docker service logs --tail 200 indra_node
