+++ modules/cf-core/src/methods/app-instance/install.spec.ts
@@ -1,8 +1,7 @@
-import { ConnextStore, getMemoryStore } from "@connext/store";
+import { getMemoryStore } from "@connext/store";
 import {
   EXPECTED_CONTRACT_NAMES_IN_NETWORK_CONTEXT,
   NetworkContext,
-  ProtocolNames,
   IStoreService,
 } from "@connext/types";
 import {
@@ -14,11 +13,10 @@ import {
 import { Wallet } from "ethers";
 import { AddressZero, HashZero, Zero } from "ethers/constants";
 import { JsonRpcProvider } from "ethers/providers";
-import { anything, instance, mock, when } from "ts-mockito";
+import { instance, mock } from "ts-mockito";
 
 import {
   NO_APP_IDENTITY_HASH_TO_INSTALL,
-  NO_MULTISIG_FOR_APP_IDENTITY_HASH,
   NO_PROPOSED_APP_INSTANCE_FOR_APP_IDENTITY_HASH,
 } from "../../errors";
 import { ProtocolRunner } from "../../machine";
@@ -71,31 +69,9 @@ describe("Can handle correct & incorrect installs", () => {
     ).rejects.toThrowError(NO_PROPOSED_APP_INSTANCE_FOR_APP_IDENTITY_HASH(HashZero));
   });
 
-  it("fails to install without the appIdentityHash being in a channel", async () => {
-    expect.hasAssertions();
-
-    const mockedStore: IStoreService = mock(ConnextStore);
-
-    const appIdentityHash = getRandomBytes32();
-    const appInstanceProposal = createAppInstanceProposalForTest(appIdentityHash);
-
-    when(mockedStore.getAppProposal(appIdentityHash)).thenResolve(appInstanceProposal);
-
-    when(mockedStore.getStateChannelByAppIdentityHash(appIdentityHash)).thenThrow(
-      Error(NO_MULTISIG_FOR_APP_IDENTITY_HASH),
-    );
-
-    await expect(
-      install(instance(mockedStore), protocolRunner, { appIdentityHash }, initiatorIdentifier),
-    ).rejects.toThrowError(NO_MULTISIG_FOR_APP_IDENTITY_HASH);
-  });
-
   it("succeeds to install a proposed AppInstance", async () => {
     const mockedProtocolRunner = mock(ProtocolRunner);
     const protocolRunner = instance(mockedProtocolRunner);
-
-    const mockedStore: IStoreService = mock(ConnextStore);
-    const store = instance(mockedStore);
 
     const appIdentityHash = getRandomBytes32();
     const multisigAddress = Wallet.createRandom().address;
@@ -120,6 +96,16 @@ describe("Can handle correct & incorrect installs", () => {
       Zero,
     );
 
+    const commitment = {
+      appIdentity: {} as any,
+      stateTimeout: toBNJson("0"),
+      appIdentityHash,
+      appStateHash: HashZero,
+      challengeRegistryAddress: AddressZero,
+      signatures: ["0x0", "0x0"],
+      versionNumber: toBNJson(1),
+    };
+
     await store.createStateChannel(
       stateChannel.toJson(),
       {
@@ -127,19 +113,21 @@ describe("Can handle correct & incorrect installs", () => {
         to: stateChannel.multisigAddress,
         value: Zero,
       },
-      {
-        appIdentity: {} as any,
-        stateTimeout: toBNJson("0"),
-        appIdentityHash,
-        appStateHash: HashZero,
-        challengeRegistryAddress: AddressZero,
-        signatures: ["0x0", "0x0"],
-        versionNumber: toBNJson(1),
-      },
+      commitment,
     );
 
     const appInstanceProposal = createAppInstanceProposalForTest(appIdentityHash);
 
+    appInstanceProposal.abiEncodings.actionEncoding = null as any; // TODO: why?
+
+    await store.createAppProposal(
+      stateChannel.multisigAddress,
+      appInstanceProposal,
+      0,
+      commitment,
+    );
+
+/*
     when(mockedStore.getAppProposal(appIdentityHash)).thenResolve(appInstanceProposal);
 
     when(mockedStore.getStateChannelByAppIdentityHash(appIdentityHash)).thenResolve(
@@ -150,6 +138,7 @@ describe("Can handle correct & incorrect installs", () => {
     // and just returns a basic <string, StateChannel> map with the
     // expected multisigAddress in it.
     when(mockedProtocolRunner.initiateProtocol(ProtocolNames.install, anything())).thenResolve();
+*/
 
     // The AppInstanceProposal that's returned is the one that was installed, which
     // is the same one as the one that was proposed
