+++ modules/store/src/index.ts
@@ -1,7 +1,8 @@
-import { IAsyncStorage, StoreTypes, IClientStore } from "@connext/types";
+import { IAsyncStorage, IClientStore } from "@connext/types";
 
 import { ConnextStore } from "./connextStore";
 import { PisaClientBackupAPI } from "./pisaClient";
+import { StoreTypes } from "./types";
 import { WrappedAsyncStorage } from "./wrappers";
 
 export const getAsyncStore = (
@@ -30,19 +31,3 @@ export const getPostgresStore = (
     StoreTypes.Postgres,
     { postgresConnectionUri: connectionUri },
   );
-
-////////////////////////////////////////
-// TODO: the following @connext/store interface is depreciated
-// remove the following exports during next breaking release
-
-export {
-  FileStorage,
-  KeyValueStorage,
-  WrappedAsyncStorage,
-  WrappedLocalStorage,
-  WrappedMemoryStorage,
-  WrappedPostgresStorage,
-} from "./wrappers";
-export { ConnextStore } from "./connextStore";
-export { PisaClientBackupAPI } from "./pisaClient";
-export { storeDefaults, storeKeys, storePaths } from "./constants";
